Input ListView In Progress
Currently, a ListView has been created, some UI and intents have also been modified. The new activity was added to the manifest, has its own layout (for the listview rows and the activity as a whole). In the future, the UI for the Input activity should be redone, and most importantly the ListView needs some sort of click listener and text modifier functionality so that the user can fill out the inputs as necessary.