Completed ListView
Completed the ListView option input page. Using information from https://demonuts.com/kotlin-listview-with-edittext/ to create modifiable editTexts instead of TextViews as in the original ListView input page. The new approach includes an EditModel class which is responsible for getting and setting the text in side of individual rows of the ListView. A populateList function dynamically initializes the contents of an ArrayList which governs the information displayed on the listView. The informational array, called editModelArrayList is responsible for storing the text to be contained in each row. In the custom adapter class, this array was passed after being populated instead of the the quantity integer, this change still allows getCount to be accessed through the .size method and also allows the modification of editTexts to be displayed. In order for the user to change text on screen and have their changes stored in the editModelArrayList to be displayed, a ChangeTextListener was implemented on the editTexts so that when their text was modified, the changes were also applied in the array and thus displayed. Some minor UI improvements were also made, including user instruction in the form of hints on the input page and choice activity color matching. Next steps include fetching logos and animating a spinner for the choice activity, along with general clean up of the app.